# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users": {
      "get": {
        "description": "List all users. Eventmanager or superuser only. Optionally filter by role.",
        "operationId": "UsersController_findAll",
        "parameters": [
          {
            "name": "role",
            "required": false,
            "in": "query",
            "description": "Filter by user role",
            "schema": {
              "enum": [
                "superuser",
                "eventmanager",
                "client",
                "osc",
                "attendee"
              ],
              "type": "string"
            }
          },
          {
            "name": "role",
            "in": "header",
            "description": "Must be: superuser | eventmanager",
            "required": true,
            "schema": {
              "type": "string",
              "example": "eventmanager"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Users list"
          }
        },
        "security": [
          {
            "role-header": []
          }
        ],
        "summary": "Get all users",
        "tags": [
          "Users"
        ]
      },
      "post": {
        "description": "Create a new user account. Eventmanager or superuser only.",
        "operationId": "UsersController_create",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "Must be: superuser | eventmanager",
            "required": true,
            "schema": {
              "type": "string",
              "example": "eventmanager"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created"
          },
          "409": {
            "description": "Email already exists"
          }
        },
        "security": [
          {
            "role-header": []
          }
        ],
        "summary": "Create a user",
        "tags": [
          "Users"
        ]
      }
    },
    "/users/stats": {
      "get": {
        "operationId": "UsersController_getStats",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "Must be: superuser | eventmanager",
            "required": true,
            "schema": {
              "type": "string",
              "example": "eventmanager"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User stats"
          }
        },
        "security": [
          {
            "role-header": []
          }
        ],
        "summary": "Get user statistics",
        "tags": [
          "Users"
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "operationId": "UsersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID (e.g. u1)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "in": "header",
            "description": "Must be: superuser | eventmanager",
            "required": true,
            "schema": {
              "type": "string",
              "example": "eventmanager"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User found"
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "role-header": []
          }
        ],
        "summary": "Get user by ID",
        "tags": [
          "Users"
        ]
      },
      "patch": {
        "operationId": "UsersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "in": "header",
            "description": "Must be: superuser | eventmanager",
            "required": true,
            "schema": {
              "type": "string",
              "example": "eventmanager"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated"
          }
        },
        "security": [
          {
            "role-header": []
          }
        ],
        "summary": "Update user (partial)",
        "tags": [
          "Users"
        ]
      },
      "delete": {
        "operationId": "UsersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "in": "header",
            "description": "Must be: superuser | eventmanager",
            "required": true,
            "schema": {
              "type": "string",
              "example": "eventmanager"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User deleted"
          }
        },
        "security": [
          {
            "role-header": []
          }
        ],
        "summary": "Delete user",
        "tags": [
          "Users"
        ]
      }
    },
...
  },
  "info": {
    "title": "WEvents API",
...
    "version": "2.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "role-header": {
        "type": "apiKey",
        "in": "header",
        "name": "role",
        "description": "User role (superuser | eventmanager | client | osc | attendee)"
      }
    },
    "schemas": {
...
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Rajesh Kumar",
            "description": "Full name of the user"
          },
          "email": {
            "type": "string",
            "example": "rajesh@gmail.com",
            "description": "Unique email address"
          },
          "role": {
            "type": "string",
            "example": "eventmanager",
            "enum": [
              "superuser",
              "eventmanager",
              "client",
              "osc",
              "attendee"
            ],
            "description": "User role"
          },
          "domain": {
            "type": "string",
            "example": "Tech Conferences",
            "description": "Domain (applicable for eventmanager and client)"
          }
        },
        "required": [
          "name",
          "email",
          "role"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Rajesh Kumar Updated"
          },
          "email": {
            "type": "string",
            "example": "rajesh.new@gmail.com"
          },
          "role": {
            "type": "string",
            "example": "attendee",
            "enum": [
              "superuser",
              "eventmanager",
              "client",
              "osc",
              "attendee"
            ]
          },
          "domain": {
            "type": "string",
            "example": "Marketing Events"
          },
          "status": {
            "type": "string",
            "example": "active",
            "enum": [
              "active",
              "inactive",
              "suspended"
            ]
          }
        }
      },
...
    }
  }
}